+++ fidget-core/src/var/mod.rs
@@ -180,10 +180,18 @@ impl VarMap {
             let Some(n) = vars.first().map(|v| v.len()) else {
                 return Ok(());
             };
-            if vars.iter().all(|v| v.len() == n) {
-                Ok(())
+            if let Some((i, v)) =
+                vars.iter().enumerate().find(|(_i, v)| v.len() != n)
+            {
+                Err(MismatchedSlices {
+                    index_a: 0,
+                    length_a: n,
+                    index_b: i,
+                    length_b: v.len(),
+                }
+                .into())
             } else {
-                Err(BulkArgError::MismatchedSlices)
+                Ok(())
             }
         }
     }
@@ -214,16 +222,30 @@ pub struct BadVarSlice {
     pub expected: usize,
 }
 
+/// Error type for mismatched slice lengths in bulk evaluation
+#[derive(thiserror::Error, Debug)]
+#[error(
+    "slice lengths are mismatched: \
+    slice {index_a} has {length_a} items; \
+    slice {index_b} has {length_b} items;"
+)]
+pub struct MismatchedSlices {
+    index_a: usize,
+    length_a: usize,
+    index_b: usize,
+    length_b: usize,
+}
+
 /// Error type for checking bulk arguments for evaluation
 #[derive(thiserror::Error, Debug)]
 pub enum BulkArgError {
     /// Variable slice length does not match expected count
     #[error(transparent)]
-    BadVarSlice(BadVarSlice),
+    BadVarSlice(#[from] BadVarSlice),
 
     /// Variable slice lengths are mismatched
-    #[error("variable slice lengths are mismatched")]
-    MismatchedSlices,
+    #[error(transparent)]
+    MismatchedSlices(#[from] MismatchedSlices),
 }
 
 /// Error type for checking tracing arguments for evaluation